pyproject.toml
[project]
name = "fishbowl"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Kyle Dougan", email = "kdougan@gmail.com"},
]
dependencies = [
    "pygame-ce>=2.5.1",
    "tomli>=2.0.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "GPL3"}


[tool.pdm]
distribution = false
